Separate redis connection classes from globals (#35)
